+++ app.js
@@ -3,40 +3,25 @@ var express = require("express");
 var path = require("path");
 var logger = require("morgan");
 let cors = require("cors");
-let cookieParser = require("cookie-parser");
-const bcrypt = require("bcryptjs");
-const User = require("./models/user");
-// const bodyParser = require("body-parser");
-
-const mongoose = require("mongoose");
 const passport = require("passport");
-const passportLocal = require("passport-local").Strategy;
 const session = require("express-session");
 
-// var indexRouter = require("./routes/index");
-// var usersRouter = require("./routes/users");
-// let testAPIRouter = require("./routes/testApi");
-// let sessionRouter = require("./routes/session");
-
-const mongoDb = "mongodb://127.0.0.1/edup_app";
-mongoose.connect(mongoDb, { useUnifiedTopology: true, useNewUrlParser: true });
-const db = mongoose.connection;
-db.on("error", console.error.bind(console, "mongo connection error"));
+var usersRouter = require("./routes/users");
+let sessionsRouter = require("./routes/sessions");
 
 var app = express();
-
-// app.use(bodyParser.json());
-// app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(logger("dev"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, "public")));
 
-// app.use(cors());
+// client Build
+app.use(express.static(path.join(__dirname, "client/build")));
 
-// app.use(bodyParser.json());
-// app.use(bodyParser.urlencoded({ extended: true }));
+app.get("http://localhost:3000", (req, res) => {
+  res.sendFile(path.join(__dirname + "/client/build/index.html"));
+});
 
 app.use(
   cors({
@@ -44,9 +29,6 @@ app.use(
     credentials: true,
   })
 );
-
-// client Build
-app.use(express.static(path.join(__dirname, "client/build")));
 
 app.use(
   session({
@@ -56,56 +38,14 @@ app.use(
   })
 );
 
-// app.use(cookieParser("secretcode"));
 app.use(passport.initialize());
 app.use(passport.session());
 
 app.use(express.urlencoded({ extended: false }));
 require("./passportConfig")(passport);
 
-app.post("/login", (req, res, next) => {
-  passport.authenticate("local", (err, user, info) => {
-    if (err) throw err;
-    if (!user) res.send("No User Exists");
-    else {
-      req.logIn(user, (err) => {
-        if (err) throw err;
-        res.send("Successfully Authenticated");
-        console.log(req.user);
-      });
-    }
-  })(req, res, next);
-});
-app.post("/register", (req, res) => {
-  User.findOne({ username: req.body.username }, async (err, doc) => {
-    if (err) throw err;
-    if (doc) res.send("User Already Exists");
-    if (!doc) {
-      const hashedPassword = await bcrypt.hash(req.body.password, 10);
-
-      const newUser = new User({
-        username: req.body.username,
-        password: hashedPassword,
-      });
-      await newUser.save();
-      res.send("User Created");
-    }
-  });
-});
-app.get("/user", (req, res) => {
-  console.log("Username");
-  console.log(req.user);
-  res.send(req.user); // The req.user stores the entire user that has been authenticated inside of it.
-});
-
-// app.use("/", indexRouter);
-// app.use("/users", usersRouter);
-// app.use("/testAPI", testAPIRouter);
-// app.use("/sessions", sessionRouter);
-
-app.get("http://localhost:3000", (req, res) => {
-  res.sendFile(path.join(__dirname + "/client/build/index.html"));
-});
+app.use("/users", usersRouter);
+app.use("/sessions", sessionsRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
