+++ controllers/sessions.js
@@ -1,0 +1,19 @@
+const passport = require("passport");
+
+const SessionsController = {
+  Create: (req, res, next) => {
+    passport.authenticate("local", (err, user, info) => {
+      if (err) throw err;
+      if (!user) res.send("No User Exists");
+      else {
+        req.logIn(user, (err) => {
+          if (err) throw err;
+          res.send("Successfully Authenticated");
+          console.log(req.user);
+        });
+      }
+    })(req, res, next);
+  },
+}
+
+module.exports = SessionsController;
