--- routes/testApi.js
@@ -1,8 +1,0 @@
-let express = require("express");
-let router = express.Router();
-
-router.get("/", function (req, res, next) {
-  res.send("API is working properly");
-});
-
-module.exports = router;
